+++ rule_estimator/rules.py
@@ -6,6 +6,8 @@ __all__ = [
     'GreaterEqualThan', 
     'LesserThan', 
     'LesserEqualThan',
+    'MultiRangeAndRule',
+    'MultiRangeOrRule'
 ]
 
 from typing import Union, List, Dict, Tuple
@@ -24,8 +26,16 @@ class CaseWhen(BusinessRule):
         if not isinstance(self.rules, list):
             self.rules = [self.rules]
 
-    def append_rule(self, new_rule:BusinessRule)->None:
-        self.rules.append(new_rule)
+    def append_rule(self, new_rule:BusinessRule, rule_id:int=None)->None:
+        if rule_id is not None:
+            rule_ids = [rule._rule_id for rule in self.rules]
+            if rule_id in rule_ids:
+                self.rules.insert(rule_ids.index(rule_id)+1, new_rule)
+            else:
+                raise ValueError(f"rule_id {rule_id} can not be found in this CaseWhen!")
+        else:
+            self.rules.append(new_rule)
+        self._stored_params['rules'] = self.rules
         
     def predict(self, X:pd.DataFrame)->np.ndarray:
         y = np.full(len(X), np.nan)
@@ -147,8 +157,9 @@ class CaseWhen(BusinessRule):
     def replace_rule(self, rule_id:int, new_rule:BusinessRule)->None:
         super().replace_rule(rule_id, new_rule)
         
-        for rule in self.rules:
-            rule.replace_rule(rule_id, new_rule)
+        if hasattr(self, "rules"):
+            for rule in self.rules:
+                rule.replace_rule(rule_id, new_rule)
             
     def get_rule_params(self, rule_id:int)->dict:
         if self._rule_id is not None and self._rule_id == rule_id:
@@ -164,6 +175,16 @@ class CaseWhen(BusinessRule):
         
         for rule in self.rules:
             rule.set_rule_params(rule_id, **params)
+
+    def _get_casewhens(self, casewhens:dict=None):
+        if self._rule_id is not None:
+            if casewhens is None:
+                casewhens =  {self._rule_id:[rule._rule_id for rule in self.rules]}
+            else:
+                casewhens[self._rule_id] = [rule._rule_id for rule in self.rules]
+        for rule in self.rules:
+            casewhens = rule._get_casewhens(casewhens)
+        return casewhens
             
     def add_to_igraph(self, graph:Graph=None)->Graph:
         graph = super().add_to_igraph(graph)
@@ -241,3 +262,87 @@ class LesserEqualThan(BusinessRule):
 
     def __rulerepr__(self):
         return f"LesserEqualThan: If {self.col} <= {self.cutoff} then predict {self.prediction}"
+
+
+class MultiRangeAndRule(BusinessRule):
+    def __init__(self, range_dict, prediction, default=None):
+        """
+        Predicts prediction if all range conditions hold for all cols in
+        range_dict. range_dict can contain multiple ranges per col.
+
+        range_dict should be of the format 
+        ```
+        range_dict = {
+            'petal length (cm)': [[4.1, 4.7], [5.2, 7.5]],
+            'petal width (cm)': [1.6, 2.6]
+        }
+        ```
+
+        """
+        super().__init__()
+    
+    def __rule__(self, X):
+        
+        def AND_masks(masks):
+            for i, mask in enumerate(masks):
+                combined_mask = mask if i == 0 else combined_mask & mask
+            return combined_mask
+        
+        def OR_masks(masks):
+            for i, mask in enumerate(masks):
+                combined_mask = mask if i == 0 else combined_mask | mask
+            return combined_mask
+        
+        def generate_range_mask(X, col, col_range):
+            if isinstance(col_range[0], list):
+                return OR_masks([(X[col] > l) & (X[col] < h) for l, h in col_range])
+            return (X[col] > col_range[0]) & (X[col] < col_range[1])
+        
+        return AND_masks([generate_range_mask(X, col, col_range) for col, col_range in self.range_dict.items()])
+        
+    def __rulerepr__(self):
+        return ("MultiRangeAndRule: If " 
+                    + " AND ".join([f"{k} in {v}" for k, v in self.range_dict.items()])
+                    + f" then predict {self.prediction}")
+
+
+class MultiRangeOrRule(BusinessRule):
+    def __init__(self, range_dict, prediction, default=None):
+        """
+        Predicts prediction if any range conditions hold for any cols in
+        range_dict. range_dict can contain multiple ranges per col.
+
+        range_dict should be of the format 
+        ```
+        range_dict = {
+            'petal length (cm)': [[4.1, 4.7], [5.2, 7.5]],
+            'petal width (cm)': [1.6, 2.6]
+        }
+        ```
+
+        """
+        super().__init__()
+    
+    def __rule__(self, X):
+        
+        def AND_masks(masks):
+            for i, mask in enumerate(masks):
+                combined_mask = mask if i == 0 else combined_mask & mask
+            return combined_mask
+        
+        def OR_masks(masks):
+            for i, mask in enumerate(masks):
+                combined_mask = mask if i == 0 else combined_mask | mask
+            return combined_mask
+        
+        def generate_range_mask(X, col, col_range):
+            if isinstance(col_range[0], list):
+                return OR_masks([(X[col] > l) & (X[col] < h) for l, h in col_range])
+            return (X[col] > col_range[0]) & (X[col] < col_range[1])
+        
+        return OR_masks([generate_range_mask(X, col, col_range) for col, col_range in self.range_dict.items()])
+        
+    def __rulerepr__(self):
+        return ("MultiRangeOrRule: If " 
+                    + " OR ".join([f"{k} in {v}" for k, v in self.range_dict.items()])
+                    + f" then predict {self.prediction}")